subje

+++ hw5/Subject.cpp
@@ -14,7 +14,7 @@ void Subject::Attach (Observer* ob) {
 }
 
 void Subject ::Detach (Observer* ob) {
-	int i=0;
+	size_t i=0;
 	for (i = 0; i < m_observers.size(); i++)
 		if (m_observers[i] == ob) break;
 	m_observers.erase(m_observers.begin() + i);
@@ -22,6 +22,6 @@ void Subject ::Detach (Observer* ob) {
 
 
 void Subject::Notify() {
-	for (int i = 0; i < m_observers.size(); i++)
+	for (size_t i = 0; i < m_observers.size(); i++)
 		(m_observers[i])->Update(this);
 }
